+++ src/main/java/com/packt/web/Dao/TaxiDaoInterface.java
@@ -6,10 +6,10 @@ import com.packt.web.bean.Taxi;
 
 public interface TaxiDaoInterface {
 	
-	public Taxi getById(int id);
+	public Taxi getById(Long id);
 	public List<Taxi> getAll();
 	public void persist(Taxi taxi);
-	public void remove(int id);
+	public void remove(Long id);
 	public void update(Taxi taxi); 
 
 }
