+++ src/main/java/com/packt/web/service/MessageServiceInterface.java
@@ -6,9 +6,9 @@ import com.packt.web.bean.Message;
 
 public interface MessageServiceInterface {
 
-	public List<Message> getMessage(int id);
-	public Message getMessageById(int id);
-	public void removeMessageByChatId(int id);
+	public List<Message> getMessage(Long id);
+	public Message getMessageById(Long id);
+	public void removeMessageByChatId(Long id);
 	public void update(Message m);
 	public Message addMessage(Message message);
